Update: Fixed int to floats

+++ HW2/sample.cpp
@@ -1,15 +1,15 @@
 #include "sample.hpp"
 
-float sampleMean(float tests, std::vector<int> &x) {
-    int sum = std::accumulate(x.begin(), x.end(), 0);
+float sampleMean(float tests, std::vector<float> &x) {
+    float sum = std::accumulate(x.begin(), x.end(), 0);
     float mean = (float) sum / tests;
     return mean;
 }
 
-float sampleVariance(float tests, float mean, std::vector<int> &x) {
+float sampleVariance(float tests, float mean, std::vector<float> &x) {
     int max = x.size();
     float acc = 0;
-    int temp;
+    float temp;
     for (int i = 0; i < max; i++) {
         temp = x[i] - mean; 
         acc += pow(temp,2);
