+++ HW2/sample.hpp
@@ -4,6 +4,6 @@
 #include <numeric>
 #include <cmath>
 
-float sampleMean (float tests, std::vector<int> &x);
+float sampleMean (float tests, std::vector<float> &x);
 
-float sampleVariance (float tests,float mean, std::vector<int> &x);
+float sampleVariance (float tests,float mean, std::vector<float> &x);
